refactor(engine): place OOM diagnostics under runtime (#244)
