+++ areno/api/trainers/ppo.py
@@ -31,7 +31,7 @@ from areno.api.dashboard import record_dashboard_state
 from areno.api.rewards import make_reward_record
 from areno.api.roles import MissingRoleCapability, ModelRole
 from areno.api.trainers.policy_only import PolicyOnlyTrainer
-from areno.engine.oom_diagnostics import OOMStage, oom_stage
+from areno.engine.runtime.oom_diagnostics import OOMStage, oom_stage
 
 logger = logging.getLogger(__name__)
