+++ areno/api/trainers/sft.py
@@ -25,7 +25,7 @@ import areno.api
 from areno.api.dashboard import record_dashboard_state
 from areno.api.data_utils import prompt_response_to_tokens_and_mask
 from areno.api.tokenizer import configure_chat_template_enable_thinking
-from areno.engine.oom_diagnostics import OOMStage, oom_stage
+from areno.engine.runtime.oom_diagnostics import OOMStage, oom_stage
 
 
 class SFTTrainer:
